+++ robolo-app/src/main/java/com/ksm/robolo/roboloapp/dto/AddressDTO.java
@@ -1,0 +1,7 @@
+package com.ksm.robolo.roboloapp.dto;
+
+/**
+ * Created by Piotrek on 10.10.2017.
+ */
+public class AddressDTO {
+}
