+++ robolo-app/src/main/java/com/ksm/robolo/roboloapp/dto/Task.java
@@ -1,4 +1,4 @@
-package com.ksm.robolo.roboloapp.models;
+package com.ksm.robolo.roboloapp.dto;
 
 /**
  * Created by Piotrek on 10.10.2017.
