--- robolo-app/src/main/java/com/ksm/robolo/roboloapp/models/Address.java
@@ -1,7 +1,0 @@
-package com.ksm.robolo.roboloapp.models;
-
-/**
- * Created by Piotrek on 10.10.2017.
- */
-public class Address {
-}
