+++ robolo-app/src/main/java/com/ksm/robolo/roboloapp/services/impl/ProjectServiceImpl.java
@@ -1,9 +1,10 @@
 package com.ksm.robolo.roboloapp.services.impl;
 
 import com.ksm.robolo.roboloapp.domain.ProjectEntity;
-import com.ksm.robolo.roboloapp.models.Project;
+import com.ksm.robolo.roboloapp.dto.Project;
 import com.ksm.robolo.roboloapp.repository.ProjectRepository;
 import com.ksm.robolo.roboloapp.services.ProjectService;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +13,8 @@ import java.util.List;
 
 @Service
 public class ProjectServiceImpl implements ProjectService{
+
+    private Logger logger = Logger.getLogger(ProjectServiceImpl.class);
 
     private ProjectRepository projectRepository;
 
@@ -22,6 +25,7 @@ public class ProjectServiceImpl implements ProjectService{
 
     @Override
     public List<Project> getAllProjects() {
+        logger.info("get all projects called");
         List<ProjectEntity> projectEntityList = projectRepository.findAll();
         List<Project> projectList = new LinkedList<>();
